[author fix]
There are a fair number of changes in this commit.

To begin with, there is a lot more defensive code to address the
confusion over whether there should be a .json / .p7s file at the end.
Some corrections atop the previous commit as well.

In addition, the code now reverses the previous loop so that we can
properly support mud files that contain only an egress or ingress ACL.

In the process of doing this, I did away with at least one useless
utility function, and changed how we count ACLs in the code.  That
might require a looksy by others.